+++ lexicon/providers/hetzner.py
@@ -169,7 +169,8 @@ class Provider(BaseProvider):
         :raises KeyError, ValueError: If the response is malformed
         :raises urllib.error.HttpError: If request to /zones did not return 200
         """
-        payload = self._get("/zones")
+        filter_obj = {"name": domain}
+        payload = self._get("/zones", filter_obj)
         zones = payload["zones"]
         for zone in zones:
             if zone["name"] == domain: